Domains: Update email forwarding behavior if site has custom domain
Currently, if you navigate to Plan -> Email from WordPress.com, you won't see
the email forwarding option since a domain isn't selected. This defaults back to
the primary domain on the site if the site has a custom domain. This establishes
a bit more consistency between the Domains -> Email and Plan -> Email flows.